requirements.txt
boto3==1.28.40
botocore==1.31.40
Flask==2.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.6.2
